Add AppManager and fix bugs
Add AppManager and fix bugs

+++ ETL/FileManager.cs
@@ -8,66 +8,81 @@ using System.Timers;
 using ETL.Models;
 
 namespace ETL
+
 {
     internal class FileManager
-    {
+    {      
+        readonly string pathOutputBase = @"..\..\..\Files\Folder_B\";
+        string pathOutputToday = "";
+        int outputCounter = 0;
 
         FolderReader folderReader = new();
-        FileSystemWatcher CSVwatcher = new(@"..\..\..\Files\Folder_A\", "*.csv");
-        FileSystemWatcher TXTwatcher = new(@"..\..\..\Files\Folder_A\", "*.txt");
-        string pathOutputBase = @"..\..\..\Files\Folder_B\";
-        string pathOutputToday = "";
-        MetaLogWriter metawriter;
-        LineConverter lineConverter = new();
-        CSVConverter csvconverter;
-        int outputCounter = 0;
+        MetaLogWorker metalogworker = new();
+        readonly FileConverter fileConverter = new();      
+        readonly System.Timers.Timer timer = new(1000);
+
         public FileManager()
         {
-            pathOutputToday = pathOutputBase + DateTime.Now.ToString("yyyy-dd-MM") + @"\";
-
-            System.Timers.Timer timer = new(1500);
             timer.Elapsed += ProcessFiles;
+        }
+
+        public void Start()
+        {
             timer.Enabled = true;
         }
 
-        private async void ProcessFiles(object? sender, ElapsedEventArgs e)
+        public void Stop()
         {
-            await ProcessDefaultFilesAsync();
+            timer.Enabled = false;
         }
 
         private async Task WriteToJsonFromOutputAsync(Output output)
         {
+            pathOutputToday = pathOutputBase + DateTime.Now.ToString("yyyy-dd-MM") + @"\";
             if (!Directory.Exists(pathOutputToday))
             {
                 Directory.CreateDirectory(pathOutputToday);
                 outputCounter = 0;
             }
-
-            using (StreamWriter sw = new(pathOutputToday + "output" + outputCounter + ".json"))
+            
+            using (StreamWriter sw = new(pathOutputToday + "output" + outputCounter++ + ".json"))
             {
+                //outputCounter++;
                 await sw.WriteAsync(JsonSerializer.Serialize(output));
-                outputCounter++;
             }
         }
-
-        private async Task ProcessDefaultFilesAsync()
+        private async void ProcessFiles(object? sender, ElapsedEventArgs e)
         {
+            await ProcessFileAsync();
+        }
+
+        private async Task ProcessFileAsync()
+        {
+
             await Task.Run(() =>
             {
-                List<string> defaultFilesPathes = folderReader.GetFilesPath().ToList();
-                List<string> csvPaths = defaultFilesPathes.Where(x => x.Contains(".csv")).ToList();
-                List<string> txtPaths = defaultFilesPathes.Where(x => x.Contains(".txt")).ToList();
-
-                csvPaths.ForEach(async x =>
+                if (DateTime.Now.Hour == 23 && DateTime.Now.Minute == 59 && DateTime.Now.Second >= 59)
                 {
-                    Output output = await lineConverter.GetOutputFromCSVFileAsync(x);
+                    metalogworker.SetPath(pathOutputBase + DateTime.Now.ToString("yyyy-dd-MM") + @$"\meta.log");
+                    metalogworker.Write();
+                    metalogworker.Dispose();
+                }
+
+                List<string> FilesPaths = folderReader.GetFilesPath().ToList();
+                if (FilesPaths.Count == 0) return;
+                List<string> csvPaths = FilesPaths.Where(x => x.Contains(".csv")).ToList();
+                List<string> txtPaths = FilesPaths.Where(x => x.Contains(".txt")).ToList();
+
+                csvPaths?.ForEach(async x =>
+                {
+                    Output output = await fileConverter.GetOutputFromCSVFileAsync(x, metalogworker);
                     await WriteToJsonFromOutputAsync(output);
                     File.Delete(x);
                 });
 
-                txtPaths.ForEach(async x =>
+                txtPaths?.ForEach(async x =>
                 {
-                    Output output = await lineConverter.GetOutputFromTXTFileAsync(x);
+                    Output output = await fileConverter.GetOutputFromTXTFileAsync(x, metalogworker);
                     await WriteToJsonFromOutputAsync(output);
                     File.Delete(x);
                 });
